+++ apps/openmw/mwmechanics/npcstats.cpp
@@ -226,9 +226,9 @@ void MWMechanics::NpcStats::useSkill (int skillIndex, const ESM::Class& class_, 
 
 void MWMechanics::NpcStats::increaseSkill(int skillIndex, const ESM::Class &class_, bool preserveProgress, bool readBook)
 {
-    int base = getSkill (skillIndex).getBase();
+    float base = getSkill (skillIndex).getBase();
 
-    if (base >= 100)
+    if (base >= 100.f)
         return;
 
     base += 1;
@@ -299,7 +299,7 @@ void MWMechanics::NpcStats::levelUp()
     for (int i=0; i<ESM::Attribute::Length; ++i)
         mSkillIncreases[i] = 0;
 
-    const int endurance = getAttribute(ESM::Attribute::Endurance).getBase();
+    const float endurance = getAttribute(ESM::Attribute::Endurance).getBase();
 
     // "When you gain a level, in addition to increasing three primary attributes, your Health
     // will automatically increase by 10% of your Endurance attribute. If you increased Endurance this level,
@@ -316,8 +316,8 @@ void MWMechanics::NpcStats::levelUp()
 
 void MWMechanics::NpcStats::updateHealth()
 {
-    const int endurance = getAttribute(ESM::Attribute::Endurance).getBase();
-    const int strength = getAttribute(ESM::Attribute::Strength).getBase();
+    const float endurance = getAttribute(ESM::Attribute::Endurance).getBase();
+    const float strength = getAttribute(ESM::Attribute::Strength).getBase();
 
     setHealth(floor(0.5f * (strength + endurance)));
 }
