+++ apps/openmw/mwmechanics/pickpocket.cpp
@@ -22,8 +22,8 @@ namespace MWMechanics
     float Pickpocket::getChanceModifier(const MWWorld::Ptr &ptr, float add)
     {
         NpcStats& stats = ptr.getClass().getNpcStats(ptr);
-        float agility = static_cast<float>(stats.getAttribute(ESM::Attribute::Agility).getModified());
-        float luck = static_cast<float>(stats.getAttribute(ESM::Attribute::Luck).getModified());
+        float agility = stats.getAttribute(ESM::Attribute::Agility).getModified();
+        float luck = stats.getAttribute(ESM::Attribute::Luck).getModified();
         float sneak = static_cast<float>(ptr.getClass().getSkill(ptr, ESM::Skill::Sneak));
         return (add + 0.2f * agility + 0.1f * luck + sneak) * stats.getFatigueTerm();
     }
